backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydub
python-multipart